+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.0.6]
+### Fixed
+- default dimensions values changed to 1
+
+### Added
+- added automaticaly weight unit change to kg
+- added settings link in plugins list
+
 ## [1.0.5]
 ### Fixed
 - fixed token value, when is empty